+++ e2e-tests/development-runtime/cypress/integration/navigation/linking.js
@@ -1,10 +1,10 @@
 describe(`navigation`, () => {
   beforeEach(() => {
-    cy.visit(`/`).waitForAPI(`onRouteUpdate`)
+    cy.visit(`/`).waitForRouteChange()
   })
 
   it(`displays content from other pages`, () => {
-    cy.visit(`/page-2`).waitForAPI(`onRouteUpdate`)
+    cy.visit(`/page-2`).waitForRouteChange()
 
     cy.getTestElement(`page-2-message`)
       .invoke(`text`)
@@ -29,7 +29,7 @@ describe(`navigation`, () => {
     beforeEach(() => {
       cy.getTestElement(`broken-link`)
         .click()
-        .waitForAPI(`onRouteUpdate`)
+        .waitForRouteChange()
     })
 
     it(`displays 404 page on broken link`, () => {