+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-node.js
@@ -10,6 +10,7 @@ jest.mock(`fs`, () => {
  * We mock sharp because it depends on fs implementation (which is mocked)
  * this causes test failures, so mock it to avoid
  */
+
 jest.mock(`sharp`, () => {
   let sharp = jest.fn(
     () =>
@@ -20,17 +21,41 @@ jest.mock(`sharp`, () => {
         toFile() {
           return Promise.resolve()
         }
+        metadata() {
+          return {
+            width: 128,
+            height: 128,
+          }
+        }
       }()
   )
+
   sharp.simd = jest.fn()
   sharp.concurrency = jest.fn()
+
   return sharp
 })
+
+jest.mock(`gatsby/dist/utils/create-content-digest`, () =>
+  jest.fn(() => `contentDigest`)
+)
 
 const fs = require(`fs`)
 const path = require(`path`)
 const sharp = require(`sharp`)
+const reporter = {
+  activityTimer: jest.fn().mockImplementation(function() {
+    return {
+      start: jest.fn(),
+      end: jest.fn(),
+    }
+  }),
+}
 const { onPostBootstrap } = require(`../gatsby-node`)
+
+const apiArgs = {
+  reporter,
+}
 
 const manifestOptions = {
   name: `GatsbyJS`,
@@ -43,6 +68,12 @@ const manifestOptions = {
     {
       src: `icons/icon-48x48.png`,
       sizes: `48x48`,
+      type: `image/png`,
+      purpose: `all`,
+    },
+    {
+      src: `icons/icon-128x128.png`,
+      sizes: `128x128`,
       type: `image/png`,
     },
   ],
@@ -60,7 +91,7 @@ describe(`Test plugin manifest options`, () => {
   })
 
   it(`correctly works with default parameters`, async () => {
-    await onPostBootstrap([], {
+    await onPostBootstrap(apiArgs, {
       name: `GatsbyJS`,
       short_name: `GatsbyJS`,
       start_url: `/`,
@@ -80,14 +111,8 @@ describe(`Test plugin manifest options`, () => {
     const icon = `pretend/this/exists.png`
     const size = 48
 
-    await onPostBootstrap([], {
-      name: `GatsbyJS`,
-      short_name: `GatsbyJS`,
-      start_url: `/`,
-      background_color: `#f7f0eb`,
-      theme_color: `#a2466c`,
-      display: `standalone`,
-      icon,
+    const pluginSpecificOptions = {
+      icon: icon,
       icons: [
         {
           src: `icons/icon-48x48.png`,
@@ -95,30 +120,27 @@ describe(`Test plugin manifest options`, () => {
           type: `image/png`,
         },
       ],
+    }
+
+    await onPostBootstrap(apiArgs, {
+      ...manifestOptions,
+      ...pluginSpecificOptions,
     })
 
     expect(sharp).toHaveBeenCalledWith(icon, { density: size })
-    expect(sharp).toHaveBeenCalledTimes(1)
+    expect(sharp).toHaveBeenCalledTimes(2)
   })
 
   it(`fails on non existing icon`, async () => {
     fs.statSync.mockReturnValueOnce({ isFile: () => false })
 
-    return onPostBootstrap([], {
-      name: `GatsbyJS`,
-      short_name: `GatsbyJS`,
-      start_url: `/`,
-      background_color: `#f7f0eb`,
-      theme_color: `#a2466c`,
-      display: `standalone`,
+    const pluginSpecificOptions = {
       icon: `non/existing/path`,
-      icons: [
-        {
-          src: `icons/icon-48x48.png`,
-          sizes: `48x48`,
-          type: `image/png`,
-        },
-      ],
+    }
+
+    return onPostBootstrap(apiArgs, {
+      ...manifestOptions,
+      ...pluginSpecificOptions,
     }).catch(err => {
       expect(sharp).toHaveBeenCalledTimes(0)
       expect(err).toBe(
@@ -134,11 +156,15 @@ describe(`Test plugin manifest options`, () => {
       plugins: [],
       theme_color_in_head: false,
       cache_busting_mode: `name`,
+      include_favicon: true,
+      crossOrigin: `anonymous`,
+      icon_options: {},
     }
-    await onPostBootstrap([], {
+    await onPostBootstrap(apiArgs, {
       ...manifestOptions,
       ...pluginSpecificOptions,
     })
+
     expect(sharp).toHaveBeenCalledTimes(0)
     const content = JSON.parse(fs.writeFileSync.mock.calls[0][1])
     expect(content).toEqual(manifestOptions)
@@ -152,12 +178,12 @@ describe(`Test plugin manifest options`, () => {
       legacy: true,
       cache_busting_mode: `name`,
     }
-    await onPostBootstrap([], {
+    await onPostBootstrap(apiArgs, {
       ...manifestOptions,
       ...pluginSpecificOptions,
     })
 
-    expect(sharp).toHaveBeenCalledTimes(1)
+    expect(sharp).toHaveBeenCalledTimes(3)
     const content = JSON.parse(fs.writeFileSync.mock.calls[0][1])
     expect(content).toEqual(manifestOptions)
   })
@@ -170,13 +196,33 @@ describe(`Test plugin manifest options`, () => {
       legacy: true,
       cache_busting_mode: `none`,
     }
-    await onPostBootstrap([], {
+    await onPostBootstrap(apiArgs, {
       ...manifestOptions,
       ...pluginSpecificOptions,
     })
 
-    expect(sharp).toHaveBeenCalledTimes(1)
+    expect(sharp).toHaveBeenCalledTimes(3)
     const content = JSON.parse(fs.writeFileSync.mock.calls[0][1])
     expect(content).toEqual(manifestOptions)
+  })
+
+  it(`icon options iterator adds options and the icon array take precedence`, async () => {
+    fs.statSync.mockReturnValueOnce({ isFile: () => true })
+
+    const pluginSpecificOptions = {
+      icon: `images/gatsby-logo.png`,
+      icon_options: {
+        purpose: `maskable`,
+      },
+    }
+    await onPostBootstrap(apiArgs, {
+      ...manifestOptions,
+      ...pluginSpecificOptions,
+    })
+
+    expect(sharp).toHaveBeenCalledTimes(3)
+    const content = JSON.parse(fs.writeFileSync.mock.calls[0][1])
+    expect(content.icons[0].purpose).toEqual(`all`)
+    expect(content.icons[1].purpose).toEqual(`maskable`)
   })
 })