+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typography/compare/gatsby-plugin-typography@2.2.10...gatsby-plugin-typography@2.2.11) (2019-04-08)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 ## [2.2.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-typography/compare/gatsby-plugin-typography@2.2.9...gatsby-plugin-typography@2.2.10) (2019-03-13)
 
 ### Bug Fixes